# Improve performance of basic incidence matrix computation