A substantial amount of time is spent in this line
https://github.com/lanl-ansi/PowerModels.jl/blob/8824212b3d2f35f94f7dd72f37ff76e3419c478f/src/core/data_basic.jl#L250
mainly because the `sort` does two dictionary lookups every time is needs to compare two branches.

This is un-necessary for two reasons:
* basic network data has branches indexed from `1` to `E` (number of branches) and no disabled branches
* the subsequent call to `sparse` will re-sort any non-ordered indices if needed

Removing this step, plus some type annotations to help the compiler, result in a substantial performance improvement.
note that relative gains get better as system size increases.
___

Performance benchmark on the `9241_pegase` case from PGLib

Before
```jl
BenchmarkTools.Trial: 91 samples with 1 evaluation per sample.
 Range (min … max):  52.836 ms … 68.059 ms  ┊ GC (min … max): 0.00% … 19.06%
 Time  (median):     54.393 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   55.223 ms ±  2.611 ms  ┊ GC (mean ± σ):  1.80% ±  3.92%

   ▁▁ ▂███▄▁▄                                                  
  ▅██████████▅▃▁▁▁▁▁▁▁▁▁▃▅▁▁▃▅▆▃▅▃▁▁▁▁▁▁▁▁▁▁▁▁▃▃▁▁▁▁▁▁▁▁▁▁▁▁▃ ▁
  52.8 ms         Histogram: frequency by time          65 ms <

 Memory estimate: 7.12 MiB, allocs estimate: 95887.
```
After
```jl
BenchmarkTools.Trial: 456 samples with 1 evaluation per sample.
 Range (min … max):  10.062 ms … 22.820 ms  ┊ GC (min … max): 0.00% … 0.00%
 Time  (median):     10.464 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   10.962 ms ±  1.485 ms  ┊ GC (mean ± σ):  2.49% ± 6.83%

  ▃▆█▅▃▂▆▂                                                     
  █████████▇▇▇▄▁▅▁▁▁▁▁▄▄▁▁▆▆▆▅▅▆▇▅▁▄▁▁▁▁▁▁▄▁▁▄▁▁▁▁▁▄▄▄▁▁▁▁▄▁▄ ▇
  10.1 ms      Histogram: log(frequency) by time        18 ms <

 Memory estimate: 2.84 MiB, allocs estimate: 32133.
```