@odow 2025-02-25
What about
```Julia
    E, N = length(data["branch"])::Int, length(data["bus"])::Int
    # I = [..., e, e, ...]
    I = repeat(1:E; inner = 2)
    J = zeros(Int, 2 * E)
    for e in 1:E
        branch = data["branch"]["$e"]
        J[2e-1] = branch["f_bus"]::Int
        J[2e] = branch["t_bus"]::Int
    end
    # V = [..., 1, -1, ...]
    V = repeat([1, -1]; inner = 2)
```