@mtanneau 2025-02-26
Thanks! This gives a nice extra 10% performance boost :)

```jl
BenchmarkTools.Trial: 517 samples with 1 evaluation per sample.
 Range (min … max):  8.979 ms … 20.571 ms  ┊ GC (min … max): 0.00% … 48.89%
 Time  (median):     9.165 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   9.671 ms ±  1.367 ms  ┊ GC (mean ± σ):  3.19% ±  8.24%

  ██▅ ▃▂▁   ▄▃                                                
  ███████▅▆▆██▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▄██▅▄▅▁▅▆▁▁▄▁▄▁▁▁▁▁▁▁▄▁▁▄▄ ▇
  8.98 ms      Histogram: log(frequency) by time     15.7 ms <

 Memory estimate: 2.84 MiB, allocs estimate: 32135.
```