@ccoffrin 2025-02-28
Updates look good to me. Just need to add a note to the change log and merge.

Open to ideas for adding a test, but also know it can be very hard to accurately capture runtime performance in CI.